pyproject.toml
[project]
name = "obsidian-semantic"
version = "0.1.0"
description = "Semantic search for Obsidian vaults"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9",
    "lancedb>=0.5",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "structlog>=24.0",
    "mcp>=1.26.0",
]

[project.optional-dependencies]
gemini = [
    "google-generativeai>=0.5",
]
dev = ["pytest>=8.0", "pytest-cov>=4.0", "ruff>=0.1"]

[project.scripts]
obsidian-semantic = "obsidian_semantic.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/obsidian_semantic"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
